svn merge -r5800:5807 https://autoplot.svn.sourceforge.net/svnroot/autoplot/autoplot/branches/autoplot2010/ .
git-svn-id: https://svn.code.sf.net/p/autoplot/code/autoplot/trunk@5808 0c594bb0-8543-0410-a25e-d8ec440261db

+++ VirboAutoplot/compile-application-all.sh
@@ -23,6 +23,13 @@ echo "\${AP_VERSION}=${AP_VERSION}"
 JAVAC=$JAVA_HOME/bin/javac
 JAR=$JAVA_HOME/bin/jar
 
+if [ \! -f $JAVAC ]; then
+   echo ""
+   echo "Can't find javac in JAVA_HOME=$JAVA_HOME"
+   echo ""
+   exit -1
+fi
+
 rm -r -f temp-src/
 mkdir temp-src/
 rm -r -f temp-classes/
